print the first 50 fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,19 +8,19 @@
 int main(void)
 {
 	int count = 2;
-	int i = 1, j = 2;
-	int k;
+	long int i = 1, j = 2;
+	long int k;
 
-	printf("%d, ", i);
-	while(count <= 50)
+	printf("%lu, ", i);
+	while (count <= 50)
 	{
-		if(count == 50)
+		if (count == 50)
 		{
-			printf("%d, ", j);
+			printf("%lu, ", j);
 		}
 		else
 		{
-			printf("%d, ", j);
+			printf("%lu, ", j);
 		}
 
 		k = j;
